Fix macos bundle icon, swap out notifier library on macos due to default action issues

+++ scripts/generate_icons.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Update the Stash icon throughout the project from a master stash-logo.png
 
-# Imagemagick, and go packages 2goarray and rsrc are required.
+# Imagemagick, and go packages 2goarray, icns and rsrc are required.
 # Copy a high-resolution stash-logo.png to this stash/scripts folder
 # and run this script from said folder, commit the result.
 
@@ -29,6 +29,13 @@ if [ ! -e "$GOPATH/bin/rsrc" ]; then
     exit
 fi
 
+if [ ! -e "$GOPATH/bin/icnsify" ]; then
+    echo "Missing Dependency:"
+    echo "Please run the following /outside/ of the stash folder:"
+    echo "go get github.com/jackmordaunt/icns/v2/cmd/icnsify@latest" 
+    exit
+fi
+
 # Favicon, used for web favicon, windows systray icon, windows executable icon
 convert stash-logo.png -define icon:auto-resize=256,64,48,32,16 favicon.ico
 cp favicon.ico ../ui/v2.5/public/
@@ -52,9 +59,8 @@ echo >> ../pkg/desktop/favicon_unix.go
 
 # MacOS, used for bundle icon
 # https://developer.apple.com/library/archive/documentation/CoreFoundation/Conceptual/CFBundles/BundleTypes/BundleTypes.html
-# "By convention, this file takes the name of the bundle and an extension of .icns; the image format can be any supported type"
-convert stash-logo.png -resize x1024 macos.png
-mv macos.png macos-bundle/Contents/Resources/stash.icns
+"$GOPATH"/bin/icnsify -i stash-logo.png -o icon.icns
+mv icon.icns macos-bundle/Contents/Resources/icon.icns
 
 # cleanup
 rm favicon.png favicon.ico
